+++ fastapi_docker/.env
@@ -1,0 +1,4 @@
+DB_CONN='postgresql://postgres:postgres@postgres_compose:5432/fastapi-kc'
+POSTGRES_USER='postgres'
+POSTGRES_PASSWORD='postgres'
+POSTGRES_DB='fastapi-kc'
