+++ fastapi_docker/docker-compose.yml
@@ -18,7 +18,7 @@ services:
     build:
       context: .
       dockerfile: Dockerfile
-    container_name: fastapi_kc
+    container_name: fastapi_kc_compose
     restart: always
     environment:
       DB_CONN: ${DB_CONN}
