+++ fastapi_docker/requirements.txt
@@ -10,3 +10,5 @@ scipy
 skops
 scikit-learn
 pydantic
+pytest
+httpx
